box demo - NOTE: had to use ps and vs version 4.0, rather than 5.0 on older desktop

+++ DX11/Core/GameTimer.cs
@@ -24,7 +24,7 @@
             _stopped = false;
 
             var countsPerSec = Stopwatch.Frequency;
-            _secondsPerCount = 1.0 / (double)countsPerSec;
+            _secondsPerCount = 1.0 / countsPerSec;
 
         }
